+++ library/route.h
@@ -15,8 +15,8 @@ public:
 	dnet_route_list(dnet_node *node);
 	~dnet_route_list();
 
-	int enable_backend(size_t backend_id, int group_id, dnet_raw_id *ids, size_t ids_count);
-	int disable_backend(size_t backend_id);
+	int enable_backend(uint32_t backend_id, int group_id, dnet_raw_id *ids, size_t ids_count);
+	int disable_backend(uint32_t backend_id);
 
 	int on_reverse_lookup(struct dnet_net_state *st, struct dnet_cmd *cmd, void *data);
 	int on_join(struct dnet_net_state *st, struct dnet_cmd *cmd, void *data);
@@ -25,7 +25,7 @@ public:
 	int send_all_ids_nolock(dnet_net_state *st, struct dnet_id *id, uint64_t trans,
 		unsigned int command, int reply, int direct);
 protected:
-	void send_update_to_states(dnet_cmd *cmd, size_t backend_id);
+	void send_update_to_states(dnet_cmd *cmd, uint32_t backend_id);
 
 private:
 	dnet_node *m_node;
@@ -52,8 +52,8 @@ typedef struct dnet_route_list_t dnet_route_list;
 dnet_route_list *dnet_route_list_create(struct dnet_node *node);
 void dnet_route_list_destroy(dnet_route_list *route);
 
-int dnet_route_list_enable_backend(dnet_route_list *route, size_t backend_id, int group_id, struct dnet_raw_id *ids, size_t ids_count);
-int dnet_route_list_disable_backend(dnet_route_list *route, size_t backend_id);
+int dnet_route_list_enable_backend(dnet_route_list *route, uint32_t backend_id, int group_id, struct dnet_raw_id *ids, size_t ids_count);
+int dnet_route_list_disable_backend(dnet_route_list *route, uint32_t backend_id);
 
 int dnet_route_list_send_all_ids_nolock(struct dnet_net_state *st, struct dnet_id *id, uint64_t trans,
 	unsigned int command, int reply, int direct);
